backend/pyproject.toml
[project]
name = "interview-panel-backend"
version = "0.1.0"
description = "Multi-Agent Interview Panel — FastAPI + LangGraph Backend"
requires-python = ">=3.11"
dependencies = [
    # --- Web Framework ---
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # --- AI / LLM ---
    "langgraph>=0.2.0",
    "langchain-google-genai>=4.0.0",
    "langchain-groq>=0.2.0",
    "langchain-core>=0.3.0",
    "nvidia-riva-client>=2.25.1",
    "openai>=1.40.0",

    # --- PDF Parsing ---
    "PyMuPDF>=1.24.0",

    # --- Database ---
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",

    # --- MCP Toolbox ---
    "toolbox-langchain>=0.1.0",

    # --- Config ---
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # --- Utilities ---
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
